fix2

+++ database/migrations/2026_04_29_134053_add_vendor_id_to_orders_table.php
@@ -12,7 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->foreignId('vendor_id')->nullable()->constrained('users')->onDelete('set null');
+            if (!Schema::hasColumn('orders', 'vendor_id'))
+                $table->foreignId('vendor_id')->nullable()->constrained('users')->onDelete('set null');
         });
     }
